requirements.txt
fugashi
unidic-lite
pywebview